requirements.txt
firebase-admin==7.2.0
google-api-python-client==2.193.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
requests==2.32.5
python-dotenv==1.0.1